+++ test/upgrade/dry_run_does_not_apply_changes_test.dart
@@ -11,7 +11,7 @@ import '../descriptor.dart' as d;
 import '../test_pub.dart';
 
 void main() {
-  test('--dry-run shows report but does not apply changes', () async {
+  test('--dry-run: shows report, changes nothing', () async {
     await servePackages((builder) {
       builder.serve('foo', '1.0.0');
       builder.serve('foo', '2.0.0');
@@ -22,61 +22,90 @@ void main() {
 
     await pubGet();
 
+    await d.dir(appPath, [
+      d.file('pubspec.lock', contains('1.0.0')),
+      d.dir('.dart_tool'),
+    ]).validate();
+
     // Change the pubspec.
     await d.appDir({'foo': 'any'}).create();
 
-    // Also delete the "packages" directory.
-    deleteEntry(path.join(d.sandbox, appPath, 'packages'));
+    // Also delete the ".dart_tool" directory.
+    deleteEntry(path.join(d.sandbox, appPath, '.dart_tool'));
 
     // Do the dry run.
     await pubUpgrade(
-        args: ['--dry-run'],
-        output: allOf([
-          contains('> foo 2.0.0 (was 1.0.0)'),
-          contains('Would change 1 dependency.')
-        ]));
+      args: ['--dry-run'],
+      output: allOf([
+        contains('> foo 2.0.0 (was 1.0.0)'),
+        contains('Would change 1 dependency.'),
+      ]),
+    );
 
     await d.dir(appPath, [
       // The lockfile should be unmodified.
       d.file('pubspec.lock', contains('1.0.0')),
-      // The "packages" directory should not have been regenerated.
-      d.nothing('packages')
+      // The ".dart_tool" directory should not have been regenerated.
+      d.nothing('.dart_tool')
     ]).validate();
   });
 
-  test(
-      '--dry-run shows report but does not apply changes even with breaking flag',
-      () async {
+  test('--dry-run --major-versions: shows report, changes nothing', () async {
     await servePackages((builder) {
       builder.serve('foo', '1.0.0');
       builder.serve('foo', '2.0.0');
     });
 
-    // Create the lockfile.
-    await d.appDir({'foo': '1.0.0'}).create();
+    await d.appDir({'foo': '^1.0.0'}).create();
 
     await pubGet();
 
-    // Recreating the appdir because the previous one only lasts for one
-    // command.
-    await d.appDir({'foo': '1.0.0'}).create();
+    await d.dir(appPath, [
+      d.file('pubspec.lock', contains('1.0.0')),
+      d.dir('.dart_tool'),
+    ]).validate();
 
-    // Also delete the "packages" directory.
-    deleteEntry(path.join(d.sandbox, appPath, 'packages'));
+    // Also delete the ".dart_tool" directory.
+    deleteEntry(path.join(d.sandbox, appPath, '.dart_tool'));
 
     // Do the dry run.
     await pubUpgrade(
-        args: ['--dry-run', '--breaking'],
-        output: allOf([
-          contains('Detected 1 potential breaking change:'),
-          contains('foo: 1.0.0 -> ^2.0.0')
-        ]));
+      args: ['--dry-run', '--major-versions'],
+      output: allOf([
+        contains('Resolving dependencies...'),
+        contains('> foo 2.0.0 (was 1.0.0)'),
+        contains('Would change 1 dependency.'),
+        contains('Would change 1 constraint in pubspec.yaml:'),
+        contains('foo: ^1.0.0 -> ^2.0.0'),
+      ]),
+    );
 
     await d.dir(appPath, [
+      // The pubspec should not be modified.
+      d.appPubspec({'foo': '^1.0.0'}),
       // The lockfile should not be modified.
+      d.file('pubspec.lock', contains('1.0.0')),
+      // The ".dart_tool" directory should not have been regenerated.
+      d.nothing('.dart_tool')
+    ]).validate();
+
+    // Try without --dry-run
+    await pubUpgrade(
+      args: ['--major-versions'],
+      output: allOf([
+        contains('Resolving dependencies...'),
+        contains('> foo 2.0.0 (was 1.0.0)'),
+        contains('Downloading foo 2.0.0...'),
+        contains('Changed 1 dependency!'),
+        contains('Changed 1 constraint in pubspec.yaml:'),
+        contains('foo: ^1.0.0 -> ^2.0.0'),
+      ]),
+    );
+
+    await d.dir(appPath, [
+      d.appPubspec({'foo': '^2.0.0'}),
       d.file('pubspec.lock', contains('2.0.0')),
-      // The "packages" directory should not have been regenerated.
-      d.nothing('packages')
+      d.dir('.dart_tool')
     ]).validate();
   });
 }
